[Enhancement] update statistics strategy on small table (#20883)
