+++ fe/fe-core/src/main/java/com/starrocks/common/Config.java
@@ -1530,13 +1530,13 @@ public class Config extends ConfigBase {
     // If the health in statistic_full_collect_interval is lower than this value,
     // choose collect sample statistics first
     @ConfField(mutable = true)
-    public static double statistic_auto_sample_ratio = 0.3;
+    public static double statistic_auto_collect_sample_threshold = 0.3;
 
     @ConfField(mutable = true)
-    public static long statistic_auto_sample_data_size = 100L * 1024 * 1024 * 1024; // 100G
+    public static long statistic_auto_collect_table_interval_size = 5L * 1024 * 1024 * 1024; // 5G
 
     @ConfField(mutable = true)
-    public static long statistic_auto_collect_interval = 3600L * 12; // unit: second, default 12h
+    public static long statistic_auto_collect_table_interval = 3600L * 12; // unit: second, default 12h
 
     /**
      * Full statistics collection max data size
